chore: rename and move modals
