+++ web/src/lib/modals/LibraryRenameModal.svelte
